% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_sema}
\alias{clean_sema}
\title{Clean sema data}
\usage{
clean_sema(input, output, rt.trim = FALSE, rt.min = 500,
  rt.threshold = 0.5)
}
\arguments{
\item{input}{The file path to the folder containing your sema data files. Important that this is the folder, not a specific file.}

\item{output}{The file path to the csv file you want to save your cleaned data in. This needs to be specifically a .csv, not a folder.}

\item{rt.trim}{Whether to remove responses that are below the rt.min threshold on reaction time. If this is TRUE, first, surveys with 
more responses below rt.min (default 500ms) than the allowed rt.threshold (default .5 or 50%) will be replaced with missing values (NA). 
Then, individual responses below the reaction time threshold from remaining surveys will be replaced with NA values.}

\item{rt.min}{The reaction time threshold below which response times are considered invalid (if rt.trim = TRUE).}

\item{rt.threshold}{The proportion of too-fast responses (faster than rt.min) before a survey is replaced with missing values (only active if rt.trim
is set to TRUE)}
}
\description{
This function collates and performs initial cleaning of raw data from the SEMA smartphone app. It saves the cleaned data to a file of your choice.
}
