% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansema.R
\docType{package}
\name{cleansema-package}
\alias{cleansema}
\alias{cleansema-package}
\title{Clean sema data}
\description{
The cleansema package was built to rapidly take folders of raw data from the SEMA smartphone application and turn
them into a single, clean dataframe, while also calculating variables that will commonly prove useful for analysis.
}
\author{
Sean C Murphy, \email{seanchrismurphy@gmail.com}

The key function you'll want to use is the \code{\link{clean_sema}} function.
}
