% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_interval}
\alias{get_interval}
\title{Get interval}
\usage{
get_interval(semadata)
}
\arguments{
\item{semadata}{An R dataframe containing both sema_id and datedlv variables (most likely, a sema dataset -after- processing with clean_sema)}
}
\description{
This function takes a sema dataset (must contain variables sema_id, timedlv and datedlv) and calculates the time interval
between successive rows. It will overwrite the interval variable created by clean_sema, but may be useful for
re-calculating intervals on datasets you have trimmed to only included surveys with responses.
}
